Control verbosity
* add option -v to have more verbose output
* control userfriendly comments by -v

Rationale: Empty results sets are normal in any query language and not
an error. Scripts want zero output here.

But empty result sets might not be expected by the user.
So having a comment stating this and suggesting changes is helpful.

Since users are trained to repeat calls with -v if they get no output,
I just added this.